Configure an Attempt for Specific Users

Override the global settings for attempts for specific users. Any unspecified setting will fall back to the global setting. This function can only be called from a server-start context.

For verifying the attempts configurations, the current configurations and all overrides can be displayed in an interactive document via show_attempts_configuration(). This is only meant for verification purposes and should not be used in the actual exam!

configure_attempts(
  user_ids,
  opens,
  closes,
  max_attempts,
  timelimit,
  grace_period
)

show_attempts_configuration()

Arguments

user_ids
a character vector of user id's, matched against the user id as determined by the authentication provider specified in exam_config().

opens, closes
the opening and closing date/time of the exam for the given users, in the format YYYY-MM-DD HH:MM Timezone (e.g., 2020-10-15 19:15 Europe/Vienna for October 15th, 2020 at 19:15 in Vienna, Austria). The exam is only accessible within this time frame. Can also be NA to either make the exam available immediately, indefinitely, or both.

max_attempts
maximum number of attempts allowed. Can also be Inf to allow for unlimited number of attempts.

timelimit
the time limit for a single attempt either as a single number in minutes or as HH:MM. Can also be Inf to give users unlimited time.

grace_period
number of seconds of "grace" period given to users before an active attempt is disabled.

Value

configure_attempts() returns the updated settings, invisibly.

show_attempts_configuration() creates an output table to show the attempts configurations. It only shows the global configuration as well as users which have different settings than the global settings.

Details

If the configuration seems to conflict with the global configuration, a warning is displayed. Conflicts are if the global configuration does specify an option, but the user-specific override explicitly disables an option (e.g., a global time-limit, but no time-limit for some users).

See also