Exam Section Configuration

Control the behavior of each individual section of an exam. Each exam contains one or more sections, which are shown either all-at-once or progressively to the user.

section_config(section, button_label, fix_order)

Arguments

section

name of the section for which to set configuration options.

button_label

the label for the button at the end of a section. Default button label is Submit answers. If progressive=FALSE, only the button in the last section will be displayed.

fix_order

logical if the section should remain fixed in order.

Details

The section configuration can be set in the metadata of the exam file and with the section_config() function. Options set via section_config() take precedence over options set in the metadata.

The last section in the exam document is assumed to be a "final remark" and displayed to the user after an attempt is finished, without possibility of submitting the section.

Sections are identified by their name. For example, a section with title # Section 1: Multiple Choice Questions would be identified by "Section 1: Multiple Choice Questions". Note that pandoc heading identifiers (anything in {} at the end of the section title) are omitted from the name.

See also