Authenticate Users Via User Interface

Present the user with a login interface before the exam begins.

ui_auth(
  users,
  username_label = "Username",
  password_label = "Password",
  title = "Login",
  button_label = "Login",
  username_empty = "Username cannot be empty.",
  password_empty = "Passwort cannot be empty.",
  unauthorized = "The username/password are incorrect."
)

Arguments

users
a data frame with user authentication information (username and hashed password). See below for details.

username_label
label for the username input field.

password_label
label for the password input field.

title
title of the dialog.

button_label
label for the login button.

username_empty, password_empty
error messages for empty username/password field.

unauthorized
error message for incorrect username/password.

Details

This authentication provider matches the given username and password against the data frame users. The data frame users must have at least two columns: username and password. The password column must contain hashed passwords created with sodium::password_store() (e.g., sodium::password_store("password"), or a SHA-256 HMAC created with digest::hmac() (digest::hmac("exam_id", "password", algo = "sha256")).

If the users data frame also contains columns display_name and/or grading, they are added to the user object (and used as described in the authentication provider documentation).

See also